+++ src/main/java/es/sralloza/choremanagementbot/models/custom/SimpleTenant.java
@@ -6,14 +6,12 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
 
-import java.util.UUID;
-
 @Data
 @Accessors(chain = true)
 @NoArgsConstructor
 @AllArgsConstructor
 public class SimpleTenant {
     @JsonProperty("tenant_id")
-    Integer tenantId;
+    Long tenantId;
     String username;
 }
